+++ src/utils/urlResolvers.ts
@@ -1,4 +1,6 @@
 import ChatSDKConfig from "../core/ChatSDKConfig";
+import LiveChatVersion from "../core/LiveChatVersion";
+import ChatAdapterProtocols from "../core/messaging/ChatAdapterProtocols";
 import libraries from "./libraries";
 
 const resolveIC3ClientUrl = (chatSDKConfig: ChatSDKConfig): string => {
@@ -13,10 +15,43 @@ const resolveIC3ClientUrl = (chatSDKConfig: ChatSDKConfig): string => {
     return libraries.getIC3ClientCDNUrl();
 };
 
+const resolveChatAdapterUrl = (chatSDKConfig: ChatSDKConfig, liveChatVersion: LiveChatVersion, protocol: string): string => {
+    const supportedChatAdapterProtocols = [ChatAdapterProtocols.ACS, ChatAdapterProtocols.IC3];
+    if (protocol && !supportedChatAdapterProtocols.includes(protocol as string)) {
+        throw new Error(`ChatAdapter for protocol ${protocol} currently not supported`);
+    }
+
+    if (protocol === ChatAdapterProtocols.ACS || liveChatVersion === LiveChatVersion.V2) {
+        if (chatSDKConfig.chatAdapterConfig && 'webChatACSAdapterCDNUrl' in chatSDKConfig.chatAdapterConfig) {
+            return chatSDKConfig.chatAdapterConfig.webChatACSAdapterCDNUrl as string;
+        }
+
+        if (chatSDKConfig.chatAdapterConfig && 'webChatACSAdapterVersion' in chatSDKConfig.chatAdapterConfig) {
+            return libraries.getACSAdapterCDNUrl(chatSDKConfig.chatAdapterConfig.webChatACSAdapterVersion);
+        }
+
+        return libraries.getACSAdapterCDNUrl();
+    } else if (protocol === ChatAdapterProtocols.IC3 || liveChatVersion === LiveChatVersion.V1) {
+        if (chatSDKConfig.chatAdapterConfig && 'webChatIC3AdapterCDNUrl' in chatSDKConfig.chatAdapterConfig) {
+            return chatSDKConfig.chatAdapterConfig.webChatIC3AdapterCDNUrl as string;
+        }
+
+        if (chatSDKConfig.chatAdapterConfig && 'webChatIC3AdapterVersion' in chatSDKConfig.chatAdapterConfig) {
+            return libraries.getIC3AdapterCDNUrl(chatSDKConfig.chatAdapterConfig.webChatIC3AdapterVersion);
+        }
+
+        return libraries.getIC3AdapterCDNUrl();
+    }
+
+    return '';
+}
+
 export default {
-    resolveIC3ClientUrl
+    resolveIC3ClientUrl,
+    resolveChatAdapterUrl
 }
 
 export {
-    resolveIC3ClientUrl
+    resolveIC3ClientUrl,
+    resolveChatAdapterUrl
 }
